fixed the end2end auth flow
I only had to add the django app for userprofile and otpauth to the INSTALLED_APPS in settings.py. It wasn't working previously because the backend wasn't registering otpauth and userprofile. Now it should prompt the user for their grinnell email and when they click "Send Code", it should take them to the initial index.tsx screen in (tabs) which is the default screen expo gives us. The 6-digit code should appear on the terminal that the backend is running on. Additionally, please use the IOS or android simulator on your laptop to run the app because without it everyone working on the code will have to change the config.ts file. This update is currently configured to run on http://127.0.0.1:8000 which just means that it'll run on the ios/android simulator so pls don't change it. Best of luck ppl :)