fix: edit listing feature
There were only 7 categories and the default was category id 8 which doesn't exist and so when the user edits the item, it gives an error. This isn't an issue with additem because that creates a new form whereas the edit item feature changes an existing form and that's why there's some inconsistency. I still changed additem.tsx just to be safe. I also commented out the sightengine stuff cause i don't wanna use my api requests during development. We'll save this for deployment