fixed item details test + shitty back button issue
I added another serializer to the backend to get the purchase requesters and the count so that we could reference it in frontend tests. We'll probably need this to show the number of purchase requesters for an item. I also removed the custom back button. Now its native ios & android back buttons but way faster