Co-authored by Alan, Joyce, Lydia
Summary of changes: Added refresh token functionality cause previously if the access token expired the user would be logged out of the account and would have to request a new otp code. Now the app automatically gets the access token through the refresh token even if the user is on the app. We also fixed the filter and sort modal formatting issue where the reset and apply button were too high. Test cases updated too. Error pages commited b/c not sure if we wanna include them but don't wanna lose them. B/c of new changes, we now use a custom api service instead of axios. This service is built on top of axios.